# docs/specs/swagger.yaml
{"openapi":"3.0.1","info":{"title":"frescos-api-grupo-2-w2","description":"frescos-api-grupo-2-w2","contact":{"name":"frescos-api-grupo-2-w2","email":"igor.cruz@mercadolivre.com"},"version":"last"},"servers":[{"url":"https://null_frescos-api-grupo-2-w2.furyapps.io","description":"Scope null"}],"paths":{"/api/v1/fresh-products/orders":{"get":{"tags":["purchase-order-controller"],"operationId":"findPurchaseOrderById","parameters":[{"name":"orderId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"put":{"tags":["purchase-order-controller"],"operationId":"updatePurchaseOrder","parameters":[{"name":"orderId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseOrderForm"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"post":{"tags":["purchase-order-controller"],"operationId":"createPurchaseOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PurchaseOrderForm"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/warehouses":{"get":{"tags":["warehouse-controller"],"operationId":"getAllWarehouses","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"post":{"tags":["warehouse-controller"],"operationId":"createWarehouse","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WarehouseForm"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/user/supervisor":{"post":{"tags":["user-controller"],"operationId":"createSupervisor","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SupervisorForm"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/v1/user/seller":{"post":{"tags":["user-controller"],"operationId":"createSeller","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SellerForm"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/v1/user/buyer":{"post":{"tags":["user-controller"],"operationId":"createBuyer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuyerForm"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/User"}}}}}}},"/api/v1/sections":{"get":{"tags":["section-controller"],"operationId":"findBatchBySection","parameters":[{"name":"sectionId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"post":{"tags":["section-controller"],"operationId":"createSection","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SectionForm"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/fresh-products":{"get":{"tags":["product-controller"],"operationId":"getAllProducts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"post":{"tags":["product-controller"],"operationId":"createProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductForm"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/fresh-products/inboundorder":{"post":{"tags":["inbound-order-controller"],"operationId":"createInboundOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InboundOrderForm"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"operationId":"ping","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/user":{"get":{"tags":["user-controller"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/fresh-products/category/list":{"get":{"tags":["product-controller"],"operationId":"getProductsByCategory","parameters":[{"name":"productType","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/fresh-products/batch/list":{"get":{"tags":["product-controller"],"operationId":"findBatchByProduct","parameters":[{"name":"productId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/fresh-products/batch/list/order":{"get":{"tags":["product-controller"],"operationId":"findBatchByProductOrder","parameters":[{"name":"productId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"order","in":"query","required":false,"schema":{"type":"string","enum":["C","F"],"default":"C"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"PurchaseOrderForm":{"required":["buyer_id","date","products","status"],"type":"object","properties":{"date":{"type":"string","format":"date"},"status":{"type":"string","enum":["CONFIRMING","SORTING","TRANSPORT","DELIVERED"]},"products":{"type":"array","items":{"$ref":"#/components/schemas/PurchaseOrderProductsForm"}},"buyer_id":{"type":"integer","format":"int64"}}},"PurchaseOrderProductsForm":{"required":["product_id","quantity"],"type":"object","properties":{"quantity":{"minimum":0,"type":"integer","format":"int32"},"product_id":{"type":"string","format":"uuid"}}},"WarehouseForm":{"required":["supervisorId"],"type":"object","properties":{"supervisorId":{"minimum":0,"type":"integer","format":"int64"}}},"SupervisorForm":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"User":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"email":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"}}},"SellerForm":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"cnpj":{"type":"string"}}},"BuyerForm":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"cpf":{"type":"string"}}},"SectionForm":{"required":["max_capacity","product_type","warehouse_id"],"type":"object","properties":{"max_capacity":{"minimum":0,"type":"integer","format":"int64"},"product_type":{"type":"string","enum":["FRESH","REFRIGERATED","FROZEN"]},"warehouse_id":{"type":"string"}}},"ProductForm":{"required":["name","seller_id","type","value"],"type":"object","properties":{"name":{"type":"string"},"type":{"type":"string","enum":["FRESH","REFRIGERATED","FROZEN"]},"value":{"minimum":0,"type":"number"},"seller_id":{"type":"integer","format":"int64"}}},"BatchForm":{"required":["batch_number","current_quantity","current_temperature","due_date","initial_quantity","manufacturing_date","manufacturing_time","minimum_temperature","product_id"],"type":"object","properties":{"batch_number":{"minimum":0,"type":"integer","format":"int64"},"product_id":{"type":"string"},"current_temperature":{"type":"number","format":"float"},"minimum_temperature":{"type":"number","format":"float"},"initial_quantity":{"minimum":0,"type":"integer","format":"int32"},"current_quantity":{"minimum":0,"type":"integer","format":"int32"},"manufacturing_date":{"type":"string","format":"date"},"manufacturing_time":{"type":"string","format":"date-time"},"due_date":{"type":"string","format":"date"}}},"InboundOrderForm":{"required":["batch_stock","order_date","order_number","section"],"type":"object","properties":{"section":{"$ref":"#/components/schemas/InboundOrderSectionForm"},"order_date":{"type":"string","format":"date"},"order_number":{"minimum":0,"type":"integer","format":"int64"},"batch_stock":{"type":"array","items":{"$ref":"#/components/schemas/BatchForm"}}}},"InboundOrderSectionForm":{"required":["section_code","warehouse_code"],"type":"object","properties":{"section_code":{"type":"string"},"warehouse_code":{"type":"string"}}}}}}